Adicionado algoritmo para encontrar o menor caminho.

+++ TextRPGAdventure/src/grafo/Aresta.java
@@ -3,21 +3,21 @@ package grafo;
 
 public class Aresta {
 
-	private Double peso;
+	private int peso;
 	private Vertice origem;
 	private Vertice destino;
 
-	public Aresta(Double peso, Vertice origem, Vertice destino){
+	public Aresta(int peso, Vertice origem, Vertice destino){
 		this.peso = peso;
 		this.origem = origem;
 		this.destino = destino;
 	}
 
-	public Double getPeso() {
+	public int getPeso() {
 		return peso;
 	}
 
-	public void setPeso(Double peso) {
+	public void setPeso(int peso) {
 		this.peso = peso;
 	}
